Refactor bitcoin stream
