+++ multy_core/src/bitcoin/bitocin_transaction_segwit.cpp
@@ -1,0 +1,16 @@
+/* Copyright 2018 by Multy.io
+ * Licensed under Multy.io license.
+ *
+ * See LICENSE for details
+ */
+
+#include "multy_core/src/bitcoin/bitcoin_transaction_segwit.h"
+
+namespace multy_core
+{
+namespace internal
+{
+
+
+} // namespace internal
+} // namespace multy_core
